requirements.txt
# Task 1
## Core
google-genai
python-dotenv
numpy
pandas

## Vector DB (하나만 선택)
pinecone
# chromadb
# faiss-cpu

# Task 
# Prototype UI (택1)
gradio

# 성능 개선/검색 최적화에 활용 가능
scikit-learn
rank-bm25

